+++ OUDS/Core/Components/Sources/_/Internal/ControlItem/ControlItemLabel.swift
@@ -47,7 +47,7 @@ struct ControlItemLabel: View {
     // MARK: - Body
 
     var body: some View {
-        HStack(alignment: .center, spacing: theme.controlItem.spaceColumnGap) {
+        HStack(alignment: .center, spacing: theme.listItem.spaceColumnGap) {
             texts()
         }
     }
@@ -55,7 +55,7 @@ struct ControlItemLabel: View {
     // MARK: - Layout Items
 
     private func texts() -> some View {
-        VStack(alignment: .leading, spacing: theme.controlItem.spaceRowGap) {
+        VStack(alignment: .leading, spacing: theme.listItem.spaceRowGap) {
             Text(layoutData.label)
                 .labelDefaultLarge(theme)
                 .multilineTextAlignment(.leading)